+++ NEWS.md
@@ -1,6 +1,7 @@
 # Version 0.0 - Release Notes
 
 * [0.0.5]
+    * ef_copert5 update (force setting to set-points) [2023-01-24]
     * archive and web update (news, readme, index) [2023-01-19] 
     * added ef_copert5 functions (still needs work; likely to replace 
       ef_emep) [2023-01-19]